Feature #2142 - Ignore empty lists in inputs
Implemented a feature to expand wild inputs. Must also set expand_input_wildcards => "true";
Also adding an acceptance test to verify that this feature is functioning as expected and as
an example for implementation. Still need to implement functionality to extract a path from
given inputs dir to the desired inputs directory if the path is included in the wild input.
Also, need to implement accaptance of an empty inputs list, and documentation needs to be
updated for this new feature and relating functionalities.

Please note: this feature is not complete, but needs to be passed on, verified and completed
before merging into master.